+++ packages/vscode-extension-vendure-helper/CHANGELOG.md
@@ -9,6 +9,16 @@ Format:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## 0.2.0
+
+### Added
+
+- "Recently picked"-Section on top displaying your last picked items
+
+### Changed
+
+- Improved error message for when the docs-fetching fails
+
 ## 0.1.0
 
 ### Added